+++ engine/flink/management/periodic/src/main/scala/pl/touk/nussknacker/engine/management/periodic/PeriodicProcessService.scala
@@ -7,11 +7,11 @@ import pl.touk.nussknacker.engine.api.deployment._
 import pl.touk.nussknacker.engine.api.process.ProcessName
 import pl.touk.nussknacker.engine.management.periodic.db.PeriodicProcessesRepository
 import pl.touk.nussknacker.engine.management.periodic.model.PeriodicProcessDeploymentStatus.{Deployed, PeriodicProcessDeploymentStatus}
-import pl.touk.nussknacker.engine.management.periodic.model.{DeploymentWithJarData, PeriodicProcessDeployment}
+import pl.touk.nussknacker.engine.management.periodic.model.{DeploymentWithJarData, PeriodicProcessDeployment, PeriodicProcessDeploymentStatus}
 import pl.touk.nussknacker.engine.management.periodic.service._
+
 import java.time.chrono.ChronoLocalDateTime
 import java.time.{Clock, LocalDateTime}
-
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.control.NonFatal
 
@@ -113,7 +113,7 @@ class PeriodicProcessService(delegateDeploymentManager: DeploymentManager,
       delegateDeploymentManager.findJobStatus(deployedProcess.periodicProcess.processVersion.processName).flatMap { state =>
         handleFinishedAction(deployedProcess, state)
           .flatMap { needsReschedule =>
-            if (needsReschedule) reschedule(deployedProcess, state) else scheduledProcessesRepository.monad.pure(()).emptyCallback
+            if (needsReschedule) reschedule(deployedProcess) else scheduledProcessesRepository.monad.pure(()).emptyCallback
           }.runWithCallbacks
       }
     }
@@ -139,7 +139,7 @@ class PeriodicProcessService(delegateDeploymentManager: DeploymentManager,
   }
 
   //Mark process as Finished and reschedule - we do it transactionally
-  private def reschedule(deployment: PeriodicProcessDeployment, state: Option[ProcessState]): RepositoryAction[Callback] = {
+  private def reschedule(deployment: PeriodicProcessDeployment): RepositoryAction[Callback] = {
     logger.info(s"Rescheduling ${deployment.display}")
     val process = deployment.periodicProcess
     for {
@@ -150,8 +150,15 @@ class PeriodicProcessService(delegateDeploymentManager: DeploymentManager,
             handleEvent(ScheduledEvent(data, firstSchedule = false))
           }.emptyCallback
         case Right(None) =>
-          logger.info(s"No next run of ${deployment.display}. Deactivating")
-          deactivateAction(process.processVersion.processName)
+          scheduledProcessesRepository.findScheduled(deployment.periodicProcess.id).flatMap { scheduledDeployments =>
+            if (scheduledDeployments.isEmpty) {
+              logger.info(s"No next run of ${deployment.display}. Deactivating")
+              deactivateAction(process.processVersion.processName)
+            } else {
+              logger.info(s"No next run of ${deployment.display} but there are still ${scheduledDeployments.size} scheduled deployments: ${scheduledDeployments.map(_.display)}")
+              scheduledProcessesRepository.monad.pure(()).emptyCallback
+            }
+          }
         case Left(error) =>
           // This case should not happen. It would mean periodic property was valid when scheduling a process
           // but was made invalid when rescheduling again.
@@ -202,15 +209,37 @@ class PeriodicProcessService(delegateDeploymentManager: DeploymentManager,
   }
 
   /**
-    * Returns latest deployment. It can be in any status (consult [[PeriodicProcessDeploymentStatus]]).
-    *
-    * For multiple schedules only the oldest one is returned.
-    */
+   * Returns latest deployment. It can be in any status (consult [[PeriodicProcessDeploymentStatus]]).
+   * For multiple schedules only single schedule is returned in the following order:
+   * <ol>
+   * <li>If there are any deployed scenarios, then the first one is returned. Please be aware that deployment of previous
+   * schedule could fail.</li>
+   * <li>If there are any failed scenarios, then the last one is returned. We want to inform user, that some deployments
+   * failed and the scenario should be rescheduled/retried manually.
+   * <li>If there are any scheduled scenarios, then the first one to be run is returned.
+   * <li>If there are any finished scenarios, then the last one is returned. It should not happen because the scenario
+   * should be deactivated earlier.
+   * </ol>
+   */
   def getLatestDeployment(processName: ProcessName): Future[Option[PeriodicProcessDeployment]] = {
     implicit val localDateOrdering: Ordering[LocalDateTime] = Ordering.by(identity[ChronoLocalDateTime[_]])
-    //TODO: so far we return only one
+
     scheduledProcessesRepository.getLatestDeploymentForEachSchedule(processName)
-      .map(_.sortBy(_.runAt).headOption).run
+      .map(_.sortBy(_.runAt)).run
+      .map { deployments =>
+        logger.debug("Found deployments: {}", deployments.map(_.display))
+
+        def first(status: PeriodicProcessDeploymentStatus) =
+          deployments.find(_.state.status == status)
+
+        def last(status: PeriodicProcessDeploymentStatus) =
+          deployments.reverse.find(_.state.status == status)
+
+        first(PeriodicProcessDeploymentStatus.Deployed)
+          .orElse(last(PeriodicProcessDeploymentStatus.Failed))
+          .orElse(first(PeriodicProcessDeploymentStatus.Scheduled))
+          .orElse(last(PeriodicProcessDeploymentStatus.Finished))
+      }
   }
 
   def deploy(deployment: PeriodicProcessDeployment): Future[Unit] = {