+++ engine/flink/management/periodic/src/test/scala/pl/touk/nussknacker/engine/management/periodic/PeriodicProcessServiceTest.scala
@@ -3,6 +3,7 @@ package pl.touk.nussknacker.engine.management.periodic
 import com.typesafe.config.{ConfigFactory, ConfigValueFactory}
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.exceptions.TestFailedException
+import org.scalatest.prop.TableDrivenPropertyChecks
 import org.scalatest.{FunSuite, Matchers, OptionValues}
 import pl.touk.nussknacker.engine.api.ProcessVersion
 import pl.touk.nussknacker.engine.api.process.ProcessName
@@ -10,6 +11,7 @@ import pl.touk.nussknacker.engine.build.EspProcessBuilder
 import pl.touk.nussknacker.engine.canonize.ProcessCanonizer
 import pl.touk.nussknacker.engine.management.FlinkStateStatus
 import pl.touk.nussknacker.engine.management.periodic.db.PeriodicProcessesRepository.createPeriodicProcessDeployment
+import pl.touk.nussknacker.engine.management.periodic.model.PeriodicProcessDeploymentStatus.PeriodicProcessDeploymentStatus
 import pl.touk.nussknacker.engine.management.periodic.model.{PeriodicProcessDeployment, PeriodicProcessDeploymentStatus}
 import pl.touk.nussknacker.engine.management.periodic.service.ProcessConfigEnricher.EnrichedProcessConfig
 import pl.touk.nussknacker.engine.management.periodic.service.{AdditionalDeploymentDataProvider, DeployedEvent, FailedEvent, FinishedEvent, PeriodicProcessEvent, PeriodicProcessListener, ProcessConfigEnricher, ScheduledEvent}
@@ -17,7 +19,7 @@ import pl.touk.nussknacker.engine.marshall.ProcessMarshaller
 import pl.touk.nussknacker.test.PatientScalaFutures
 
 import java.time.temporal.ChronoField
-import java.time.{Clock, LocalDate}
+import java.time.{Clock, LocalDate, LocalDateTime}
 import scala.collection.mutable.ArrayBuffer
 import scala.concurrent.Future
 
@@ -25,7 +27,8 @@ class PeriodicProcessServiceTest extends FunSuite
   with Matchers
   with OptionValues
   with ScalaFutures
-  with PatientScalaFutures {
+  with PatientScalaFutures
+  with TableDrivenPropertyChecks {
 
   import org.scalatest.LoneElement._
   import pl.touk.nussknacker.engine.spel.Implicits.asSpelExpression
@@ -33,6 +36,9 @@ class PeriodicProcessServiceTest extends FunSuite
   import scala.concurrent.ExecutionContext.Implicits.global
 
   private val processName = ProcessName("test")
+  private val yearNow = LocalDate.now().get(ChronoField.YEAR)
+  private val cronInFuture = CronScheduleProperty(s"0 0 6 6 9 ? ${yearNow + 1}")
+  private val cronInPast = CronScheduleProperty(s"0 0 6 6 9 ? ${yearNow - 1}")
   private val processJson = ProcessMarshaller.toJson(ProcessCanonizer.canonize(
     EspProcessBuilder
       .id(processName.value)
@@ -103,6 +109,34 @@ class PeriodicProcessServiceTest extends FunSuite
     f.events.toList shouldBe List(FinishedEvent(finished, f.delegateDeploymentManagerStub.jobStatus), ScheduledEvent(scheduled, firstSchedule = false))
   }
 
+  test("handleFinished - should deactivate process if there are no future schedules") {
+    val f = new Fixture
+    f.repository.addActiveProcess(processName, PeriodicProcessDeploymentStatus.Deployed, scheduleProperty = cronInPast)
+    f.delegateDeploymentManagerStub.setStateStatus(FlinkStateStatus.Finished)
+
+    f.periodicProcessService.handleFinished.futureValue
+
+    val processEntity = f.repository.processEntities.loneElement
+    processEntity.active shouldBe false
+    f.repository.deploymentEntities.loneElement.status shouldBe PeriodicProcessDeploymentStatus.Finished
+    // TODO: active should be false
+    val event = createPeriodicProcessDeployment(processEntity.copy(active = true), f.repository.deploymentEntities.loneElement)
+    f.events.loneElement shouldBe FinishedEvent(event, f.delegateDeploymentManagerStub.jobStatus)
+  }
+
+  test("handleFinished - should not deactivate process if there is future schedule") {
+    val f = new Fixture
+    val scheduleProperty = MultipleScheduleProperty(Map("schedule1" -> cronInPast, "schedule2" -> cronInFuture))
+    val periodicProcessId = f.repository.addOnlyProcess(processName, scheduleProperty)
+    f.repository.addOnlyDeployment(periodicProcessId, status = PeriodicProcessDeploymentStatus.Deployed, scheduleName = Some("schedule1"))
+    f.repository.addOnlyDeployment(periodicProcessId, status = PeriodicProcessDeploymentStatus.Scheduled, scheduleName = Some("schedule2"))
+
+    f.periodicProcessService.handleFinished.futureValue
+
+    f.repository.processEntities.loneElement.active shouldBe true
+    f.repository.deploymentEntities.map(_.status) should contain only (PeriodicProcessDeploymentStatus.Finished, PeriodicProcessDeploymentStatus.Scheduled)
+  }
+
   test("handle first schedule") {
     val f = new Fixture
 
@@ -166,9 +200,6 @@ class PeriodicProcessServiceTest extends FunSuite
 
   test("Schedule new scenario only if at least one date in the future") {
     val f = new Fixture
-    val yearNow = LocalDate.now().get(ChronoField.YEAR)
-    val cronInFuture = CronScheduleProperty(s"0 0 6 6 9 ? ${yearNow + 1}")
-    val cronInPast = CronScheduleProperty(s"0 0 6 6 9 ? ${yearNow - 1}")
 
     def tryToSchedule(schedule: ScheduleProperty): Unit = f.periodicProcessService.schedule(schedule, ProcessVersion.empty, processJson).futureValue
 
@@ -177,5 +208,35 @@ class PeriodicProcessServiceTest extends FunSuite
 
     intercept[TestFailedException](tryToSchedule(cronInPast)).getCause shouldBe a[PeriodicProcessException]
     intercept[TestFailedException](tryToSchedule(MultipleScheduleProperty(Map("s1" -> cronInPast, "s2" -> cronInPast)))).getCause shouldBe a[PeriodicProcessException]
+  }
+
+  test("getLatestDeployment - should return correct deployment for multiple schedules") {
+    val schedules@(schedule1 :: schedule2 :: Nil) = List("schedule1", "schedule2")
+    val table = Table(
+      ("statuses", "expected status", "expected schedule name"),
+      (List(PeriodicProcessDeploymentStatus.Scheduled, PeriodicProcessDeploymentStatus.Scheduled), PeriodicProcessDeploymentStatus.Scheduled, schedule1),
+      (List(PeriodicProcessDeploymentStatus.Deployed, PeriodicProcessDeploymentStatus.Scheduled), PeriodicProcessDeploymentStatus.Deployed, schedule1),
+      (List(PeriodicProcessDeploymentStatus.Finished, PeriodicProcessDeploymentStatus.Scheduled), PeriodicProcessDeploymentStatus.Scheduled, schedule2),
+      (List(PeriodicProcessDeploymentStatus.Finished, PeriodicProcessDeploymentStatus.Finished), PeriodicProcessDeploymentStatus.Finished, schedule2),
+      (List(PeriodicProcessDeploymentStatus.Failed, PeriodicProcessDeploymentStatus.Scheduled), PeriodicProcessDeploymentStatus.Failed, schedule1),
+      (List(PeriodicProcessDeploymentStatus.Failed, PeriodicProcessDeploymentStatus.Deployed), PeriodicProcessDeploymentStatus.Deployed, schedule2),
+      (List(PeriodicProcessDeploymentStatus.Failed, PeriodicProcessDeploymentStatus.Finished), PeriodicProcessDeploymentStatus.Failed, schedule1),
+      (List(PeriodicProcessDeploymentStatus.Failed, PeriodicProcessDeploymentStatus.Failed), PeriodicProcessDeploymentStatus.Failed, schedule2),
+    )
+
+    forAll(table) { (statuses: List[PeriodicProcessDeploymentStatus], expectedStatus: PeriodicProcessDeploymentStatus, expectedScheduleName: String) =>
+      val f = new Fixture
+      val now = LocalDateTime.now()
+      val scheduleProperty = MultipleScheduleProperty(Map("schedule1" -> cronInPast, "schedule2" -> cronInFuture))
+      val periodicProcessId = f.repository.addOnlyProcess(processName, scheduleProperty)
+      statuses.zip(schedules).zipWithIndex.foreach { case ((status, schedule), index) =>
+        f.repository.addOnlyDeployment(periodicProcessId, status = status, runAt = now.plusDays(index), scheduleName = Some(schedule))
+      }
+
+      val deployment = f.periodicProcessService.getLatestDeployment(processName).futureValue.value
+
+      deployment.state.status shouldBe expectedStatus
+      deployment.scheduleName shouldBe Some(expectedScheduleName)
+    }
   }
 }